[REFACTOR] #10 Stateful 변수들StateHolde에 저장
