+++ app/src/main/java/com/example/android_accountbook_13/presenter/AccountBookApp.kt
@@ -2,31 +2,23 @@ package com.example.android_accountbook_13.presenter
 
 import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.compose.currentBackStackEntryAsState
-import androidx.navigation.compose.rememberNavController
 import com.example.android_accountbook_13.R
 import com.example.android_accountbook_13.presenter.component.AccountBookBottomAppBar
 import com.example.android_accountbook_13.presenter.component.AccountBookTopAppBar
 import com.example.android_accountbook_13.presenter.navigation.AccountBookNavHost
-import com.example.android_accountbook_13.presenter.navigation.History
-import com.example.android_accountbook_13.presenter.navigation.bottomTabScreens
-import com.example.android_accountbook_13.presenter.navigation.navigateSingleTopTo
 import com.example.android_accountbook_13.ui.theme.MyTheme
 
 @Composable
-fun MainScreen() {
-    MyTheme() {
-        val navController = rememberNavController()
-        val currentBackStack by navController.currentBackStackEntryAsState()
-        val currentDestination = currentBackStack?.destination
-        val currentScreen = bottomTabScreens.find { it.route == currentDestination?.route }
+fun AccountBookApp() {
+    MyTheme {
+        val appState = rememberAccountBookAppState()
 
         Scaffold(
+            scaffoldState = appState.scaffoldState,
             topBar = {
                 AccountBookTopAppBar(
-                    title = "2022년 7월",
+                    title = "",
                     leftVectorResource = R.drawable.ic_left,
                     rightVectorResource = R.drawable.ic_right,
                     onLeftClick = {},
@@ -35,15 +27,15 @@ fun MainScreen() {
             },
             bottomBar = {
                 AccountBookBottomAppBar(
-                    destination = currentScreen ?: History,
+                    destination = appState.currentScreen ?: History,
                     onClick = { newScreen ->
-                        navController.navigateSingleTopTo(newScreen.route)
+                        appState.navController.navigateSingleTopTo(newScreen.route)
                     }
                 )
             }
         ) {
             AccountBookNavHost(
-                navController,
+                appState.navController,
                 History.route
             )
         }
@@ -53,5 +45,5 @@ fun MainScreen() {
 @Composable
 @Preview(showBackground = true)
 fun MainScreenPreview() {
-    MainScreen()
+    AccountBookApp()
 }
