+++ app/src/main/java/com/example/android_accountbook_13/presenter/MainActivity.kt
@@ -3,9 +3,6 @@ package com.example.android_accountbook_13.presenter
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.*
-import androidx.compose.ui.tooling.preview.Preview
-import com.example.android_accountbook_13.ui.theme.MyTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -13,17 +10,7 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            MyTheme {
-                MainScreen()
-            }
+            AccountBookApp()
         }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun DefaultPreview() {
-    MyTheme {
-        MainScreen()
     }
 }
