+++ app/src/main/java/com/example/android_accountbook_13/presenter/navigation/AccountBookNavHost.kt
@@ -1,10 +1,13 @@
 package com.example.android_accountbook_13.presenter.navigation
 
 import androidx.compose.runtime.Composable
-import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import com.example.android_accountbook_13.presenter.Calendar
+import com.example.android_accountbook_13.presenter.History
+import com.example.android_accountbook_13.presenter.Setting
+import com.example.android_accountbook_13.presenter.Statistic
 import com.example.android_accountbook_13.presenter.calendar.CalendarScreen
 import com.example.android_accountbook_13.presenter.history.HistoryScreen
 import com.example.android_accountbook_13.presenter.setting.SettingScreen
@@ -34,17 +37,5 @@ fun AccountBookNavHost(
         composable(Setting.route) {
             SettingScreen()
         }
-    }
-}
-
-fun NavHostController.navigateSingleTopTo(route: String) {
-    this.navigate(route) {
-        popUpTo(
-            this@navigateSingleTopTo.graph.findStartDestination().id
-        ) {
-            saveState = true
-        }
-        launchSingleTop = true
-        restoreState = true
     }
 }
